initial commit for typescript conversion
