+++ package.json
@@ -24,13 +24,16 @@
     "url": "https://github.com/arnaudNYC/mov-to-gif/issues"
   },
   "homepage": "https://github.com/arnaudNYC/mov-to-gif#readme",
-  "dependencies": {},
   "devDependencies": {
+    "@types/node": "^18.11.18",
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.1",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-prettier": "^3.1.2",
-    "prettier": "^1.19.1"
+    "prettier": "^1.19.1",
+    "typescript": "^4.9.4"
   }
 }
